+++ src/test/java/be/ugent/idlab/knows/dataio/records/JSONRecordTest.java
@@ -9,17 +9,17 @@ import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class JSONRecordTest {
 
     @Test
     public void testTwoDotsIterator() throws Exception{
-        SourceIterator it = new JSONSourceIterator(new LocalFileAccess("artworks.json", "src/test/resources/json/jsonPathTwoDotsTest", NAMESPACES.XSD + "string"), "$");
-        List<String> results = new ArrayList<>();
-        it.forEachRemaining(r -> {
-            results.add(r.get("..children[*].id").toString());
-        });
-        assertTrue(results.contains("[[29,91,31,2815,177,95,97,273,544,2653,451,799]]"));
+        List<Object> results;
+        try (SourceIterator it = new JSONSourceIterator(new LocalFileAccess("artworks.json", "src/test/resources/json/jsonPathTwoDotsTest", NAMESPACES.XSD + "string"), "$")) {
+            results = new ArrayList<>();
+            it.forEachRemaining(r -> results.addAll((List<?>)r.get("..children[*].id").getValue()));
+        }
+        assertEquals(List.of(29L,91L,31L,2815L,177L,95L,97L,273L,544L,2653L,451L,799L), results);
     }
 }